pyproject.toml

[tool.poetry]
name = "ai-travel-agent"
version = "0.1.1"
description = ""
authors = ["Nir Bar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.0"
langchain-google-genai = "^2.0.0"
google-genai = "^0.8.0"
langgraph = "^0.2.0"
grandalf = "^0.8"
serpapi = "^0.1.5"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
langgraph-cli = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
